# swagger.json
{"info":{"title":"dawpy server","description":"This is the daw server api","version":"0.1.0"},"paths":{"/":{"get":{"summary":"Get Root","description":"get current daw state","operationId":"get_root__get","responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Daw"}}},"parameters":[],"produces":["application/json"]}},"/daw":{"get":{"summary":"Get Daw","description":"get current daw state","operationId":"get_daw_daw_get","responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Daw"}}},"parameters":[],"produces":["application/json"]}},"/daw/render":{"get":{"summary":"Render Playlist","description":"render current playlist","operationId":"render_playlist_daw_render_get","responses":{"200":{"description":"Successful Response","schema":{"title":"Response Render Playlist Daw Render Get","type":"string"}}},"parameters":[],"produces":["application/json"]}},"/playlist/load/{playlist_name}":{"get":{"summary":"Save Playlist","description":"save current daw state","operationId":"save_playlist_playlist_load__playlist_name__get","parameters":[{"required":true,"name":"playlist_name","in":"path","type":"string"}],"responses":{"200":{"description":"Successful Response","schema":{}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"produces":["application/json"]}},"/playlist/save":{"get":{"summary":"Save Playlist","description":"save current daw state","operationId":"save_playlist_playlist_save_get","responses":{"200":{"description":"Successful Response","schema":{"title":"Response Save Playlist Playlist Save Get","type":"string"}}},"parameters":[],"produces":["application/json"]}},"/playlist":{"get":{"summary":"Get Current Playlist","description":"get current playlist","operationId":"get_current_playlist_playlist_get","responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Playlist"}}},"parameters":[],"produces":["application/json"]},"post":{"summary":"Create Playlist","description":"create a new playlist","operationId":"create_playlist_playlist_post","responses":{"200":{"description":"Successful Response","schema":{}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"parameters":[{"required":true,"name":"body","in":"body","schema":{"$ref":"#/definitions/Playlist"}}],"consumes":["application/json"],"produces":["application/json"]}},"/playlist/add/{index}/{pattern_name}":{"post":{"summary":"Add Pattern To Song Map","description":"add pattern to songmap at index","operationId":"add_pattern_to_song_map_playlist_add__index___pattern_name__post","parameters":[{"required":true,"name":"pattern_name","in":"path","type":"string"},{"required":true,"name":"index","in":"path","type":"integer"}],"responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Pattern"}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"produces":["application/json"]}},"/pattern":{"post":{"summary":"Create Pattern","description":"create a new pattern","operationId":"create_pattern_pattern_post","responses":{"200":{"description":"Successful Response","schema":{}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"parameters":[{"required":true,"name":"body","in":"body","schema":{"$ref":"#/definitions/Pattern"}}],"consumes":["application/json"],"produces":["application/json"]}},"/pattern/{pattern_name}":{"get":{"summary":"Get Pattern By Name","description":"Get pattern by name","operationId":"get_pattern_by_name_pattern__pattern_name__get","parameters":[{"required":true,"name":"pattern_name","in":"path","type":"string"}],"responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Pattern"}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"produces":["application/json"]},"put":{"summary":"Update Pattern","description":"update a pattern by name","operationId":"update_pattern_pattern__pattern_name__put","parameters":[{"required":true,"name":"pattern_name","in":"path","type":"string"},{"required":true,"name":"body","in":"body","schema":{"$ref":"#/definitions/Pattern"}}],"responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Pattern"}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"consumes":["application/json"],"produces":["application/json"]},"patch":{"summary":"Configure Pattern","description":"Configure Pattern","operationId":"configure_pattern_pattern__pattern_name__patch","parameters":[{"required":true,"name":"pattern_name","in":"path","type":"string"}],"responses":{"200":{"description":"Successful Response","schema":{}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"produces":["application/json"]}},"/plugin":{"get":{"summary":"Get Plugins","description":"get registered plugins","operationId":"get_plugins_plugin_get","responses":{"200":{"description":"Successful Response","schema":{"title":"Response Get Plugins Plugin Get","type":"array","items":{"$ref":"#/definitions/Plugin"}}}},"parameters":[],"produces":["application/json"]},"post":{"summary":"Register Plugin","description":"register a new plugin","operationId":"register_plugin_plugin_post","responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Daw"}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"parameters":[{"required":true,"name":"body","in":"body","schema":{"$ref":"#/definitions/Plugin"}}],"consumes":["application/json"],"produces":["application/json"]}},"/plugin/{plugin_name}":{"get":{"summary":"Get Plugin By Name","description":"Get plugin by name","operationId":"get_plugin_by_name_plugin__plugin_name__get","parameters":[{"required":true,"name":"plugin_name","in":"path","type":"string"}],"responses":{"200":{"description":"Successful Response","schema":{"$ref":"#/definitions/Plugin"}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"produces":["application/json"]},"patch":{"summary":"Configure Plugin","description":"Configure Plugin","operationId":"configure_plugin_plugin__plugin_name__patch","parameters":[{"required":true,"name":"plugin_name","in":"path","type":"string"}],"responses":{"200":{"description":"Successful Response","schema":{}},"422":{"description":"Validation Error","schema":{"$ref":"#/definitions/HTTPValidationError"}}},"produces":["application/json"]}}},"swagger":"2.0","definitions":{"Daw":{"title":"Daw","type":"object","properties":{"daw_config_folder":{"title":"Daw Config Folder","type":"string"},"daw_config":{"$ref":"#/definitions/DawConfig"},"playlist":{"$ref":"#/definitions/Playlist"},"registered_plugins":{"title":"Registered Plugins","type":"array","items":{"$ref":"#/definitions/Plugin"}}}},"DawConfig":{"title":"DawConfig","type":"object","properties":{"config_name":{"title":"Config Name","type":"string"},"daw_config_path":{"title":"Daw Config Path","type":"string"},"mrs_watson_64bit_path":{"title":"Mrs Watson 64Bit Path","type":"string"},"mrs_watson_32bit_path":{"title":"Mrs Watson 32Bit Path","type":"string"},"nano_host_64bit_path":{"title":"Nano Host 64Bit Path","type":"string"},"nano_host_32bit_path":{"title":"Nano Host 32Bit Path","type":"string"},"pattern_path":{"title":"Pattern Path","type":"string"},"plugin_path":{"title":"Plugin Path","type":"string"},"preset_path":{"title":"Preset Path","type":"string"},"projects_path":{"title":"Projects Path","type":"string"},"rendered_path":{"title":"Rendered Path","type":"string"}}},"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/definitions/ValidationError"}}}},"Pattern":{"title":"Pattern","type":"object","properties":{"name":{"title":"Name","type":"string"},"plugin":{"$ref":"#/definitions/Plugin"},"midi_path":{"title":"Midi Path","type":"string"}}},"Playlist":{"title":"Playlist","required":["name","bpm"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"bpm":{"title":"Bpm","type":"integer"},"patterns":{"title":"Patterns","type":"array","items":{"$ref":"#/definitions/Pattern"}},"playlist_dict":{"title":"Playlist Dict","type":"object","additionalProperties":{"$ref":"#/definitions/Pattern"}}}},"Plugin":{"title":"Plugin","type":"object","properties":{"name":{"title":"Name","type":"string"},"dll_path":{"title":"Dll Path","type":"string"},"preset_path":{"title":"Preset Path","type":"string"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"type":"string"}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}},"x-components":{}}